refactor(publish-metrics): extract adot translators into their own file
